pyproject.toml
[tool.poetry]
name = "elfextract"
version = "0.1.0"
description = "Extract function raw binaries from ELF files"
authors = ["Michael Liao <michael.hliao@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyelftools = "^0.30"

[tool.poetry.scripts]
elfextract = 'elfextract.main:main'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
